Dropping seeds!
I can click and drop the seeds wherever I want! Which is not exactly what we want but oh well! It works! Also I fixed that stupid bug error that I was getting before. I just had to set the terrain index or something to 0 or 1 or some non negative number. Don't remember exactly.